product CRUD : ID 타입 Long으로 변경

+++ src/main/java/com/sparta/productorder/controller/ProductController.java
@@ -27,7 +27,7 @@ public class ProductController {
     }
 
     @GetMapping("/{id}")
-    public List<ProductResponseDto> getProductById(@PathVariable Integer id){
+    public List<ProductResponseDto> getProductById(@PathVariable Long id){
         return productService.getProductById(id);
     }
